feat: Add support for refreshing node and rule metrics
This commit adds support for refreshing node and rule metrics in the `cmgrImpl` struct. The `QueryNodeMetrics` and `QueryRuleMetrics` methods now have an additional `refresh` parameter, which when set to `true`, refreshes the cache and retrieves the latest data.

The `QueryNodeMetrics` method now checks if `refresh` is `true` and updates the cache with the latest node metrics if needed. It then calls the `QueryNodeMetric` method to retrieve the requested metrics.

The `QueryRuleMetrics` method follows a similar pattern, refreshing the cache with the latest rule metrics if `refresh` is `true`, and then calling the `QueryRuleMetric` method to retrieve the metrics.

This change improves the accuracy of the metrics data by allowing users to retrieve the latest metrics when needed.

Refactor code to use GetAddrHost instead of ToString for Node struct
save work
save
save work
support parse rule metric
refine metrics label
feat: better node metrics chart (#370)
chore: Update dependencies and refactor transporter package for improved performance and readability (#369)
chore: use cgo-free sqlite driver
feat: add timepicker to dashboard (#367)